Require lp8 modules, don’t import them.
Use `lp8.require` rather than `lp8.import`, to work around an issue with
`import` where it errors out because the module is already imported. I
couldn’t find whence this problem was coming, but who but me cares. It’s
not like we were using the extra functionality of `import` anyway.