Explicitly write quantity key to item config
Previously, quantities displayed were handled at dialog runtime, since most items did not
have a quantity= key unless they picked up a second item with the same ID. This change makes
every item have a quantity key, and also fixes a bug where two duplicate items would still
be listed with a quantity of 1 due to the value fallback being 0.

An additional change is the addition of the [take_item] quantity= key, which allows specifying
item quantity when it is picked up. This defaults to 1.