Improved implementation of APPLY_EXTRA_AMLA_EFFECT
This utilizes a flag in unit.variables to prevent the event from firing multiple times.
We also remove the event from the current scenario. Sadly, due to the events being kept
in the character's unit types instead of individual unit configs (for future compatibility
with new amlas), I believe this will result in the events also appearing in subsequent
scenarios, even if they won't fire